+++ app/(user)/profile/texting-compliance/register/page.js
@@ -9,4 +9,4 @@ export const metadata = meta
 
 export default function Page() {
   return <TextingComplianceRegisterPage />
-} 
+}
