+++ package.json
@@ -72,6 +72,7 @@
     "react-csv": "^2.2.2",
     "react-dom": "^19.1.0",
     "react-drag-reorder": "^1.2.0",
+    "react-google-autocomplete": "^2.7.5",
     "react-icons": "^5.5.0",
     "react-image-crop": "^10.1.5",
     "react-inlinesvg": "^4.2.0",
